fix: fix db setup in tests (#458)
* feat: initially add seed and migration files

* feat: place seeds and migrations in folders

* feat: remove not used inserts

* fix: fix adoption code test

* fix: remove init mariadb script from gh ci

* feat: add node env check

* fix: use migration both in tests and in docker compose

* refactor: use sql connection as object

* feat: separate dev and test dbs

* fix: rename migration script

* fix: fix docker compose file

+++ test/setup.ts
@@ -1,7 +1,10 @@
+import config from 'config';
+import Bluebird from 'bluebird';
 import chai from 'chai';
 import nock from 'nock';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import type { Knex } from 'knex';
 
 import { populateMemList } from '../src/lib/geoip/whitelist.js';
 import {
@@ -10,24 +13,41 @@ import {
 	populateIpRangeList,
 	populateNockCitiesList,
 } from './utils/populate-static-files.js';
-
 import chaiOas from './plugins/oas/index.js';
 import { getRedisClient, initRedis } from '../src/lib/redis/client.js';
-import { client } from '../src/lib/sql/client.js';
-import { USERS_TABLE } from '../src/lib/adopted-probes.js';
+import { client as sql } from '../src/lib/sql/client.js';
+
+const dbConfig = config.get<{ connection: { database: string, host: string } }>('db');
+
+if (!dbConfig.connection.database.endsWith('-test') && dbConfig.connection.host !== 'localhost') {
+	throw new Error(`Database name for test env needs to end with "-test" or the host must be "localhost". Got "${dbConfig.connection.database}"@"${dbConfig.connection.host}".`);
+}
 
 before(async () => {
 	chai.use(await chaiOas({ specPath: path.join(fileURLToPath(new URL('.', import.meta.url)), '../public/v1/spec.yaml') }));
+
 	await initRedis();
 	const redis = getRedisClient();
 	await redis.flushDb();
-	await client(USERS_TABLE).insert({ id: '89da69bd-a236-4ab7-9c5d-b5f52ce09959', github: 'jimaek' }).onConflict().ignore();
+
+	await dropAllTables(sql);
+	await sql.migrate.latest();
+	await sql.seed.run();
 
 	nock.disableNetConnect();
 	nock.enableNetConnect('127.0.0.1');
+
 	await populateIpList();
 	await populateDomainList();
 	await populateIpRangeList();
 	await populateMemList();
 	await populateNockCitiesList();
 });
+
+const dropAllTables = async (sql: Knex) => {
+	const allTables = (await sql('information_schema.tables')
+		.whereRaw(`table_schema = database()`)
+		.select(`table_name as table`)
+	).map(({ table }: { table: string }) => table);
+	await Bluebird.map(allTables, table => sql.schema.raw(`drop table \`${table}\``));
+};
